refactor: make GarageView class reponsible only for elements creation and rendering
