+++ async-race/src/components/Paginator.ts
@@ -1,21 +1,24 @@
-import GarageView from '../views/GarageView';
 import Communicator from './Communicator';
 import { CARS_PER_PAGE } from '../utils/constants';
+import GarageHandler from './GarageHandler';
+import GarageView from '../views/GarageView';
 
 export default class Paginator {
   communicator = new Communicator();
 
   currentPage = 1;
 
+  handler = new GarageHandler();
+
   async prevPage() {
     const cars = await this.communicator
       .getCars(GarageView.getInstance().pageCount -= 1, CARS_PER_PAGE);
-    GarageView.getInstance().updatePage(cars);
+    this.handler.updatePage(cars);
   }
 
   async nextPage() {
     const cars = await this.communicator
       .getCars(GarageView.getInstance().pageCount += 1, CARS_PER_PAGE);
-    GarageView.getInstance().updatePage(cars);
+    this.handler.updatePage(cars);
   }
 }
