+++ async-race/src/views/GarageView.ts
@@ -1,24 +1,21 @@
 /* eslint-disable no-param-reassign */
-import Communicator from '../components/Communicator';
 import Paginator from '../components/Paginator';
 import { CarType } from '../utils/types';
 import View from './View';
-import { CARS_PER_PAGE } from '../utils/constants';
+import GarageHandler from '../components/GarageHandler';
 
 export default class GarageView extends View {
   static Instance: GarageView;
 
   paginator = new Paginator();
 
-  communicator = new Communicator();
+  handler = new GarageHandler();
 
   carsWrapper = this.createCarsWrapper();
 
   viewE = this.createView();
 
-  selectedCar: HTMLElement | null = null;
-
-  totalItems = 0;
+  itemsCount = 0;
 
   pageCount = 1;
 
@@ -55,7 +52,7 @@ export default class GarageView extends View {
 
     const creationBtn = this.createElement('button', 'creation-btn');
     creationBtn.textContent = 'Create';
-    creationBtn.addEventListener('click', async () => this.handleCarAdding(carName, carColor));
+    creationBtn.addEventListener('click', async () => this.handler.handleCarAdding(carName, carColor));
 
     carCreator.append(carName, carColor, creationBtn);
     return carCreator;
@@ -71,7 +68,7 @@ export default class GarageView extends View {
 
     const updateBtn = this.createElement('button', 'update-btn');
     updateBtn.textContent = 'Update';
-    updateBtn.addEventListener('click', () => this.handleCarUpdating(carName, carColor));
+    updateBtn.addEventListener('click', () => this.handler.handleCarUpdating(carName, carColor));
 
     carUpdater.append(carName, carColor, updateBtn);
     return carUpdater;
@@ -101,49 +98,9 @@ export default class GarageView extends View {
     return this.createElement('div', 'cars-wrapper');
   }
 
-  async stuffCarsWrapper() {
-    const cars = await this.communicator.getCars(this.pageCount, CARS_PER_PAGE);
-    cars.forEach((car: CarType) => this.appendCar(this.createCarE(car)));
-  }
-
-  async updatePage(cars: CarType[]) {
-    await this.replaceCars(cars);
-    this.updatePageCounter();
-    this.checkPaginationButtons();
-    this.removeFromSelected(this.selectedCar);
-  }
-
-  async replaceCars(cars: CarType[]) {
+  replaceCars(cars: CarType[]) {
     const newCars = cars.map((car) => this.createCarE(car));
     this.carsWrapper.replaceChildren(...newCars);
-  }
-
-  updatePageCounter() {
-    (this.viewE.querySelector('.page-num') as HTMLElement)
-      .textContent = this.pageCount.toString();
-  }
-
-  checkPaginationButtons() {
-    const prevBtn = this.viewE.querySelector('.prev-btn') as HTMLElement;
-    if (this.pageCount < 2) {
-      prevBtn.classList.add('btn-disabled');
-    } else {
-      prevBtn.classList.remove('btn-disabled');
-    }
-
-    const nextBtn = this.viewE.querySelector('.next-btn') as HTMLElement;
-    if (this.totalItems <= this.pageCount * CARS_PER_PAGE) {
-      nextBtn.classList.add('btn-disabled');
-    } else {
-      nextBtn.classList.remove('btn-disabled');
-    }
-  }
-
-  async updateItemsCounter() {
-    const totalItems = await this.communicator.getTotalItems() as string;
-    (this.viewE.querySelector('.total-counter') as HTMLElement).textContent = totalItems;
-    this.totalItems = +totalItems;
-    this.checkPaginationButtons();
   }
 
   createCarE(car: CarType) {
@@ -169,9 +126,9 @@ export default class GarageView extends View {
       </div>
     </div>`);
     const removeBtn = carE.querySelector('.car-remove');
-    removeBtn?.addEventListener('click', () => this.handleCarRemoving(carE as HTMLElement));
+    removeBtn?.addEventListener('click', () => this.handler.handleCarRemoving(carE as HTMLElement));
     const selectBtn = carE.querySelector('.car-select');
-    selectBtn?.addEventListener('click', () => this.handleCarSelection(carE as HTMLElement));
+    selectBtn?.addEventListener('click', () => this.handler.handleCarSelection(carE as HTMLElement));
     return carE;
   }
 
@@ -190,94 +147,5 @@ export default class GarageView extends View {
     </g>
   </svg>
     `;
-  }
-
-  async handleCarAdding(carNameE: HTMLInputElement, carColorE: HTMLInputElement) {
-    const name = carNameE.value;
-    if (name.trim()) {
-      const color = carColorE.value;
-      const params = {
-        name,
-        color,
-      };
-
-      const car = await this.communicator.addCar(params);
-      if (this.carsWrapper.childNodes.length < CARS_PER_PAGE) {
-        this.appendCar(this.createCarE(car));
-      }
-
-      carNameE.value = '';
-      carColorE.value = '#000000';
-
-      await this.updateItemsCounter();
-    } else {
-      alert('Please enter some name for a car');
-    }
-  }
-
-  async handleCarRemoving(carE: HTMLElement) {
-    this.removeFromSelected(carE);
-    await this.communicator.removeCar(carE.dataset.id as string);
-    await this.updateItemsCounter();
-    const newCars = await this.communicator.getCars(this.pageCount, CARS_PER_PAGE);
-    const lastNewCar = newCars[CARS_PER_PAGE - 1];
-    carE.remove();
-    if (lastNewCar) this.appendCar(this.createCarE(lastNewCar));
-  }
-
-  async handleCarSelection(carE: HTMLElement) {
-    if (this.selectedCar === carE) {
-      this.removeFromSelected(carE);
-    } else {
-      this.removeFromSelected(this.selectedCar);
-      this.addToSelected(carE);
-    }
-  }
-
-  addToSelected(carE: HTMLElement) {
-    const { carsUpdater, carName, carColor } = this.getUpdatingControls();
-    carsUpdater?.classList.remove('disabled');
-
-    carName.value = carE.querySelector('.car-name')?.textContent as string;
-    carColor.value = carE.querySelector('.car-model svg')?.getAttribute('fill') as string;
-    this.selectedCar = carE;
-    carE.classList.add('selected');
-  }
-
-  removeFromSelected(carE: HTMLElement | null) {
-    const { carsUpdater, carName, carColor } = this.getUpdatingControls();
-    carsUpdater?.classList.add('disabled');
-    if (this.selectedCar === carE) {
-      this.selectedCar = null;
-      carName.value = '';
-      carColor.value = '#000000';
-    }
-    carE?.classList.remove('selected');
-  }
-
-  getUpdatingControls() {
-    const carsUpdater = this.viewE.querySelector('.cars-updater');
-    return {
-      carsUpdater,
-      carName: carsUpdater?.querySelector('.car-name') as HTMLInputElement,
-      carColor: carsUpdater?.querySelector('.car-color') as HTMLInputElement,
-    };
-  }
-
-  async handleCarUpdating(
-    carNameE: HTMLInputElement,
-    carColorE: HTMLInputElement,
-  ) {
-    const name = carNameE.value;
-    if (name.trim()) {
-      const color = carColorE.value;
-      this.communicator.updateCar(this.selectedCar?.dataset.id as string, { name, color });
-
-      (this.selectedCar?.querySelector('.car-name') as HTMLElement).textContent = name;
-      this.selectedCar?.querySelector('.car-model svg')?.setAttribute('fill', color);
-      this.removeFromSelected(this.selectedCar);
-    } else {
-      alert('Please enter some name for a car');
-    }
   }
 }
